fix timeout

+++ GA-Framework/src/problems/qbf/solvers/GA_QBF_SC.java
@@ -54,7 +54,7 @@ public class GA_QBF_SC extends GA_QBF {
 	 * 
 	 */
 	public static void main(String[] args) throws IOException {
-        Integer timeoutInSeconds = 60;
+        Long timeoutInSeconds = 60L;
 		long startTime = System.currentTimeMillis();
 		GA_QBF_SC ga = new GA_QBF_SC(10000, 100, 1.0 / 100.0, "GA-Framework/instances/qbf-sc/instance_7.txt", timeoutInSeconds);
 		Solution<Integer> bestSol = ga.solve();
